Converted Edict obj to WDict, Kanjidic obj to KDict.
Further renaming is strongly discouraged.
The names of these objects no longer refer to the source datafiles they
currently use.  I want to add various loader functions for different file
formats, and if we do so then the old names become misnomers.  Not to mention,
this eliminates confusion between the objects and the data files they're based
off of.