Changes in dictionary handling.  KRAD/RADKFILE reading implemented.
All dictionary files have been moved to the dicts subdirectory.
This will likely be the permanent location of the dict files on Windows
systems, but on Linux they'll likely be elsewhere - I'll adjust as necessary.

Paths to the dictionary files are provided by the Preferences object.  Defaults
are provided if nothing is specified in the loaded config file.

KRADFILE and RADKFILE readers have been implemented, and all data is now stored
in KDict.  I have not yet decided how I will use the stroke count data in
RADKFILE, but I am storing it nonetheless.