Makefile changes.  Slight GUI reorganization.  Handling of user's home directory.
Makefile changed:
* Windows "release" builds have symbol tables stripped.  Other
  platforms leave it intact.
* "install" and "uninstall" targets now are designated as
  "non-Windows" - on Windows they just display a message saying
  they're not supported.

GUI reorganization:
* FrameMainGUI's constructor code was reorganized so widgets are
  added to the tab controls from the lowest children upwards.
* The tabMajor and tabMinor pointers are -both- initialized to NULL.
* The word dictionary tab is now the default instead of the kanji
  dictionary.

Home directory handling:
The program now queries %AppData% or $HOME, depending on the system,
and the J-Ben config file is written to that directory.  If these
variables are for some reason not set, an error will be logged and
the file will be written to the current directory instead.